requirements.txt
python-dotenv==1.1.0
pydantic-settings==2.7.1
sqlalchemy==2.0.36
fastapi==0.115.6
uvicorn[standard]==0.34.0
discord.py==2.4.0
apscheduler==3.10.4
httpx==0.28.1
loguru==0.7.3
